agents/requirements.txt
crewai==0.51.0
crewai-tools==0.8.3
supabase==2.5.3
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.8.2
apscheduler==3.10.4
openai==1.37.0
